app/requirements.txt
asgiref==3.7.2
channels==4.0.0
daphne==4.0.0
Django==4.2.6
Django-Verify-Email==2.0.3
django-online-users
python-dotenv==1.0.0
mysqlclient==2.1.1
whitenoise==6.5.0